mission: submit week8 mission

+++ src/main/java/javalab/umc7th_mission/study/repository/MemberRepository/MemberRepository.java
@@ -3,6 +3,6 @@ package javalab.umc7th_mission.study.repository.MemberRepository;
 import javalab.umc7th_mission.study.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<Member, Integer>, MemberRepositoryCustom {
+public interface MemberRepository extends JpaRepository<Member, Long>, MemberRepositoryCustom {
 
 }
